+++ include/legacy/input.h
@@ -16,4 +16,4 @@ extern char right_click[2];
 void init_input();
 void send_keyboard_input(char wVk, DWORD dwFlags);
 Vector2<float> mouse_position();
-void move_mouse_to(int x, int y);
+void move_mouse_to(float x, float y);
